requirements.txt
joblib
sklearn
pandas
Flask
flask-httpauth
scikit-learn
Werkzeug
pydantic[email]
pymongo[srv]
python-dotenv
certifi
